# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LLI Client Location Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  companyCode: string;
  companyName: string;
  city: string;
  state: string;
  pinCode: string;
  latitude: number;
  longitude: number;
  fieldPerson: string;
  status: string;
  createdAt?: string;
  locality?: string;
  address?: string;
  fullAddress?: string;
  geoStatus?: string;
}

export interface ClientWithDistance {
  id: number;
  companyCode: string;
  companyName: string;
  city: string;
  state: string;
  pinCode: string;
  latitude: number;
  longitude: number;
  fieldPerson: string;
  status: string;
  createdAt?: string;
  locality?: string;
  address?: string;
  fullAddress?: string;
  geoStatus?: string;
  distanceKm: number;
}

export interface CityCount {
  city: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface StateCount {
  state: string;
  count: number;
}

export interface LocalityCount {
  locality: string;
  count: number;
}

export interface ClientStats {
  total: number;
  byCity: CityCount[];
  byStatus: StatusCount[];
  byState: StateCount[];
  byLocality: LocalityCount[];
}

export type ListClientsParams = {
  /**
   * Filter by city name (case-insensitive partial match)
   */
  city?: string;
  /**
   * Filter by pin code
   */
  pinCode?: string;
  /**
   * Filter by client status
   */
  status?: string;
  /**
   * Filter by locality (case-insensitive partial match)
   */
  locality?: string;
};

export type GetNearbyClientsParams = {
  /**
   * Latitude of the reference location
   */
  lat: number;
  /**
   * Longitude of the reference location
   */
  lng: number;
  /**
   * Radius in kilometers (default 10)
   */
  radius?: number;
};