# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LLI Client Location Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  companyCode: string;
  companyName: string;
  city: string;
  state: string;
  pinCode: string;
  latitude: number;
  longitude: number;
  fieldPerson: string;
  status: string;
  createdAt?: string;
  locality?: string;
  address?: string;
  fullAddress?: string;
  geoStatus?: string;
}