# lib/api-zod/src/generated/types/localityCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LLI Client Location Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface LocalityCount {
  locality: string;
  count: number;
}